security without refresh token

+++ app/src/main/java/control/self/igor/dailywisdom/controller/frontend/abstraction/AbstractCrudController.java
@@ -9,6 +9,9 @@ import org.jboss.logging.Logger;
 import org.jboss.logging.Logger.Level;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,7 +23,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.fasterxml.jackson.annotation.JsonView;
 
 import control.self.igor.dailywisdom.entity.Identifiable;
+import control.self.igor.dailywisdom.entity.UserRole;
 import control.self.igor.dailywisdom.exception.BadRequestException;
+import control.self.igor.dailywisdom.exception.ForbiddenException;
 import control.self.igor.dailywisdom.exception.NotFoundException;
 import control.self.igor.dailywisdom.json.View;
 import control.self.igor.dailywisdom.model.api.EntityCounter;
@@ -68,6 +73,9 @@ public abstract class AbstractCrudController<Entity extends Identifiable> {
 
     @PostMapping("")
     public Response createEntity(@Valid @RequestBody Entity entity) {
+	if (!isAdmin()) {
+	    throw ForbiddenException.createAdminOnlyException();
+	}
 	long id;
 	try {
 	    id = crudService.createEntity(entity);
@@ -80,6 +88,9 @@ public abstract class AbstractCrudController<Entity extends Identifiable> {
 
     @PutMapping("/{id}")
     public Response updateEntity(@PathVariable("id") long id, @Valid @RequestBody Entity entity) {
+	if (!isAdmin()) {
+	    throw ForbiddenException.createAdminOnlyException();
+	}
 	if (!crudService.updateEntity(entity)) {
 	    throw new NotFoundException();
 	}
@@ -90,6 +101,9 @@ public abstract class AbstractCrudController<Entity extends Identifiable> {
     public Response deleteEntity(@PathVariable("id") long id) {
 	if (id < 1) {
 	    throw new BadRequestException();
+	}
+	if (!isAdmin()) {
+	    throw ForbiddenException.createAdminOnlyException();
 	}
 	try {
 	    crudService.deleteEntity(id);
@@ -104,6 +118,19 @@ public abstract class AbstractCrudController<Entity extends Identifiable> {
     @GetMapping("/list/count")
     public EntityCounter countEntities() {
 	return new EntityCounter(crudService.countEntities());
+    }
+
+    protected boolean isAdmin() {
+	Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	if (authentication.getAuthorities().isEmpty()) {
+	    return false;
+	}
+	for (GrantedAuthority authority : authentication.getAuthorities()) {
+	    if (UserRole.Role.isAdmin(authority.getAuthority())) {
+		return true;
+	    }
+	}
+	return false;
     }
 
 }
